code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py37_0
astor=0.7.1=py_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
binutils_impl_linux-64=2.33.1=h53a641e_8
binutils_linux-64=2.33.1=h9595d00_16
blas=1.0=mkl
bleach=3.1.0=pypi_0
bzip2=1.0.8=h516909a_2
c-ares=1.15.0=h516909a_1001
ca-certificates=2019.11.28=hecc5488_0
cairo=1.16.0=h18b612c_1001
certifi=2019.11.28=py37hc8dfbb8_1
cffi=1.13.2=py37h2e261b9_0
cloudpickle=1.3.0=py_0
cudatoolkit=10.0.130=0
cudnn=7.6.4=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py_2
cython=0.29.15=pypi_0
cytoolz=0.10.1=py37h7b6447c_0
dask-core=2.9.0=py_0
dbus=1.13.6=he372182_0
decorator=4.4.1=py_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
expat=2.2.6=he6710b0_0
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.9.1=h8a8886c_1
gast=0.2.2=py_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_16
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.2.1=h516909a_2
glib=2.58.3=py37h6f030ca_1002
gmp=6.2.0=he1b5a44_2
gnutls=3.6.5=hd3a4fd2_1002
google-pasta=0.1.8=py_0
graphite2=1.3.13=hf484d3e_1000
grpcio=1.23.0=py37he9ae1f9_0
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
gxx_impl_linux-64=7.3.0=hdf63c60_1
gxx_linux-64=7.3.0=h553295d_16
h5py=2.10.0=nompi_py37h513d04c_100
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.5=nompi_h3c11f04_1104
hilbertcurve=1.0.3=pypi_0
icu=58.2=hf484d3e_1000
imageio=2.6.1=py37_0
importlib-metadata=1.3.0=pypi_0
intel-openmp=2019.4=243
ipykernel=5.1.3=pypi_0
ipython=7.11.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.15.2=pypi_0
jinja2=2.10.3=pypi_0
joblib=0.14.0=py_0
jpeg=9c=h14c3975_1001
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=5.3.4=pypi_0
jupyter-console=6.0.0=pypi_0
jupyter-core=4.6.1=pypi_0
keras=2.3.1=py37_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.1.0=py37hc9558a2_0
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.33.1=h53a641e_8
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libgpuarray=0.7.6=h14c3975_1003
libiconv=1.15=h516909a_1005
liblapack=3.8.0=14_mkl
liblapacke=3.8.0=14_mkl
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.1=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h56121f0_5
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=hea5a465_1
mako=1.1.0=py_0
markdown=3.1.1=py_0
markupsafe=1.1.1=py37h516909a_0
matplotlib=3.1.1=py37h5429711_0
mistune=0.8.4=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
more-itertools=8.0.2=pypi_0
multicoretsne=0.1=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=4.4.0=pypi_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_0
nibabel=3.1.0=pypi_0
ninja=1.9.0=py37hfd86e86_0
notebook=6.0.2=pypi_0
numpy=1.17.4=py37hc1035e2_0
numpy-base=1.17.4=py37hde5b4d6_0
olefile=0.46=py37_0
opencv=4.1.1=py37ha799480_1
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1d=h516909a_0
opt_einsum=3.1.0=py_0
packaging=20.4=pypi_0
pandas=1.0.1=py37h0573a6f_0
pandocfilters=1.4.2=pypi_0
parso=0.5.2=pypi_0
pcre=8.43=he1b5a44_0
pexpect=4.7.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
pixman=0.38.0=h516909a_1003
pot=0.6.0=pypi_0
prometheus-client=0.7.1=pypi_0
prompt-toolkit=2.0.10=pypi_0
protobuf=3.11.1=py37he1b5a44_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=pypi_0
pycparser=2.19=py37_0
pygments=2.5.2=pypi_0
pygpu=0.7.6=py37hc1659b7_1000
pymorton=1.0.5=pypi_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37hcca6a23_4
pyrsistent=0.15.6=pypi_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.3.1=cuda100py37h53c1284_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37h7b6447c_0
pyyaml=5.3=py37h516909a_0
pyzmq=18.1.1=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=4.6.0=pypi_0
readline=7.0=h7b6447c_5
scikit-image=0.15.0=py37he6710b0_0
scikit-learn=0.21.3=py37hd81dba3_0
scipy=1.3.1=py37h7c811a0_0
send2trash=1.5.0=pypi_0
setuptools=42.0.2=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tabulate=0.8.7=pypi_0
tensorboard=1.15.0=pyhb230dea_0
tensorboardx=1.9=py_0
tensorflow=1.15.0=gpu_py37h0f0df58_0
tensorflow-base=1.15.0=gpu_py37h9dcbed7_0
tensorflow-estimator=1.15.1=pyh2649769_0
tensorflow-gpu=1.15.0=h0d30ee6_0
termcolor=1.1.0=py_2
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
theano=1.0.4=py37he1b5a44_1001
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.4.2=py37_cu100
tornado=6.0.3=py37h516909a_0
tqdm=4.40.0=py_0
traitlets=4.3.3=pypi_0
tsne-mp=0.1.13=pypi_0
wcwidth=0.1.7=pypi_0
webencodings=0.5.1=py_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.11.2=py37h516909a_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.2.2=h516909a_1
zipp=0.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0